Python 3 support using single codebase approach. All tests are passing for pythons 2.6, 2.7, 3.2 and pypy 1.8.
For simplicity I replaced some 'iteritems' with 'items' and 'xrange' with 'range' when it doesn't matter in practice whenever iterator or list version is used in python 2.
The only place where code becomes worse is a metaclass hack, but there only 2 'hackish' lines and it seems a reasonable price for python 3 support.