sqlalchemy session actions
Remove force_flush OptionDefault and replace
with sqlalchemy_session_action which has two
valid values 'flush' and 'commit'. On create
call we preform the specified action on the session.

If None is specified, we do nothing.

If a value other than 'create', 'flush', or None is
speficied, we raise a TypeError