Clarify iterator recycling.
When using @factory.iterator, the values from the underlying generator
will be saved in memory, and reused once the initial generator has
reached its end.

This means that, if the generator relies on external information (e.g
datetime.now()), generated values won't be recomputed when that external
information is changed.

To avoid this, users should use the ``factory.Iterator(func,
    cycle=False)`` pattern, and design the target function for never
ending.

Related to issue #344.