[refactor] Require kwargs for PostGeneration
PostGenerationMethodCall could receive many positional arguments.
This did not play well with the ability to override the value at build
time: would the overridden value match one of the parameters? Should we
pass a tuple?

Instead, we force the user to use a *single* positional argument; all
others should be keyword arguments.

The clumsiness of those combined features makes it unlikely that anyone
would have been using them.

ERROR: API for PostGenerationMethodCall has changed
WARNING: It is no longer possible to use a PostGenerationMethodCall with
         a method accepting only positional arguments (i.e ``*args``).