Add RelatedFactoryList class for one-to-many support
The RelatedFactoryList class allows us to define a declaration which
returns a list of RelatedFactoryList(s).

This is different than the current RelatedFactory, which only generates
1 related object when it's parent Factory is 'created'.

This is useful for one-to-many relations. For instance, let's pretend we
have a ``Student`` model which has *many* ``Exams``. We don't want to
strictly generate *one* exam per student in our test suite, we want
to cover the case of many exams per student.

So, we'd tell the ``Student.exams`` declaration to be a
``RelatedFactoryList``, and could either tell it to generate ``length``
exams per student, or provide a ``lambda`` which would lazily get
evaluated; this allows for a dynamic number of related objects
being generated for each parent object.

This lambda can be used to create deterministic counts of related
objects per parent object, or to randomly generate "n" related
objects, where "n" is a random int. (See added test coverage
for the former case, and latter case).

This has been brought up in a few StackOverflow posts. I wasn't
satisfied so I opened this PR!

http://stackoverflow.com/questions/27120685/one-to-many-relationships-in-factory-boy/
http://stackoverflow.com/questions/14589362/creating-multiple-objects-with-foreign-key/