Evaluate extra kwargs passed to declarations.
Unroll a declaration's context before evaluating it (i.e evaluate all
lazy declarations in that context before passing them as kwargs to the
actual declaration's evaluate() / call() function).

Note: this allows for the following code:

class SomeFactory:
    value__low = factory.Sequence(lambda n: n)
    value = factory.fuzzy.FuzzyInteger(low=3, high=100)

Whereas passing ``FuzzyInteger(low=factory.Sequence(lambda n: n))``
wouldn't work as expected.

Whether that feature is supported or not has not been decided yet; it's
considered to be an undefined behaviour and a side-effect of internal
changes.

Closes #466.