Re-raise IntegrityError when django_get_or_create fails
When:
* several fields are specified in `django_get_or_create`, and
* user specifies a subset of kwargs, and
* other kwargs listed in `django_get_or_create` have values
  generated by the factory,
a `get_or_create` lookup may trigger an `IntegrityError`. Indeed, a
generated value that was not part of the original lookup is passed to
`get_or_create()`, causing a failed `get()` and leading to a `create()`
with more than user specified kwargs and triggering an `IntegrityError`.

The `IntegrityError` is handled by factory_boy, and the model is looked
up based on user specified kwargs. When the lookup fails,
factory_boy raised a `ValueError`. Re-raise the original
`IntegrityError` instead to align behavior with Django.

Fixes #606
Refs #345