Change post_generation example to look like a staticmethod
Based on the [reference docs](https://factoryboy.readthedocs.io/en/latest/reference.html?highlight=post_generation#factory.post_generation), the `post_generation` decorator doesn't behave like an instance method on the factory. The first parameter is listed as `self` in the many-to-many example, but `self` is actually a model instance.

In the other areas where `post_generation` is referenced, the first parameter seems to be named `obj`. This documentation change updates the example to be consistent with other usage.