Expose memleak in SQLAlchemyModelFactory.
The code responsible for `get_or_create` places a copy of the passed-in
params on the factory class itself; which is kept after the instance has
been created.

Those fields are part of the BuildStep context, and should be fetched
from there instead.

The issue also occurs in DjangoModelFactory.

See #781 for details.