Apply sphinxcontrib-spelling to the docs for spell checking
Project URL: https://github.com/sphinx-contrib/spelling

sphinxcontrib-spelling is a mature Sphinx extension to do spell checking
across docs. It is used by other big name projects such as Django. This
tool will now run as part of CI.

All spelling mistakes caught by the tool have been fixed. The language
is set to en_US so the British spelling of behaviour was converted to
the American spelling.

The credits.rst file is excluded from checking as it is mostly a list of
people's names which often don't appear in the dictionary.