Add support for Python 3.9
The readme was updated not to mention the Python version explicitly, it
already links to the stable branches on the Python devguide for
supported version. That’s reduces the number of elements to upgrade for
next Python versions.

Use the latest Python version for all auxiliary builds (docs, example
test, etc). That allows using the latest features in these builds.