Migrate CI from Travis to GitHub actions
GitHub actions produces faster results than Travis, resulting in faster
feedback for contributors.

The linkcheck environment was not ported to GitHub actions. This
environment consistently fails with the error:

    429 Client Error: too many requests for url: https://github.com/…

This sometimes happens in Travis, but is more frequent now with the
increase in performance. See issue #771.

The "basepython" tox configuration option was dropped as GitHub actions
default to using the latest Python version which is now 3.9. This avoid
additional configuration in both actions and tox when a new Python is
released.

Fixes #707