Deprecate factory.use_strategy()
The purpose of this function is unclear. It is not used in the library
and not really tested.

Follow the Zen of Python, keep only one way of specifying a factory
default strategy: through its options. Users can define the strategy
when calling the factory with the dedicated methods (e.g.
`MyFactory.build()`, or `MyFactory.create()`).

Reviewed-by: Raphaël Barrois