Perform context unrolling inside declarations
Unrolling the extra context for a declaration might depend on the
declaration's internals; that's typically the case with factory.Maybe:
the inner declarations might depend on the actual declaration used.

This adds `evaluate_pre` and a `evaluate_post` entrypoints to
declarations, more readable with regard to which build phase they are
used in.

Each of those will perform unrolling before calling the semi-public
actual function entrypoint (self.evaluate() for evaluate_pre,
self.call() for evaluate_post).

As a side effect, this fixes the issues with factory.Faker() when called
inside a factory.Maybe().

Closes #785 #786 #787 #788 #790 #796.