fix(fuzzy-decimal): Convert to float
Convert high and low to float explicitly, before passing them to
Random().uniform() as it does a calculation there with float values. If
a Decimal is passed in as argument (for example from a Django decimal or
money field), then this will fail.

This isn't the best solution, but it's the least invasive.

The better would be to reimplement Random().uniform as
factory.random.decimal_uniform and cast everything to Decimals. This
would retain input accuracy, but tests monkey patching uniform() would
have to be changed and this reimplementation must be kept in line with
its counterpart in the Python Standard Library.