Format code with Black
Black is a very popular Python code formatter. It ensures a consistent
coding style across the library. It is opinionated and not configurable,
so that projects adopting Black are similar.

The main advantage a code formatter is to avoid concerns about coding
style. Many text editors have an integration for Black and can be
configured to format on save.

The isort profile has been set to "black" for compatibility.
https://pycqa.github.io/isort/docs/configuration/black_compatibility/

https://github.com/psf/black

Short multiline comprehensions were kept on multiple lines using #
fmt: skip to improve the readability.