Remove confusing return in post_generation example
The return value of post generation hooks is stored for advanced uses
(i.e. saving instances for DjangoModelFactory after the post generation
hooks completed), but this use is undocumented and not part of the
public API.

After issue #316 is completed, the save on post generation for
DjangoModelFactory will be removed. That will also drop the last use for
the `_after_postgeneration()` extension point. However, the extension
point is kept to avoid disruption for projects that may be relying on
it.