Cache pip dependencies across lint and test runs
Speeds up tests, and nicer to PyPI bandwidth.

Caches unused for 7 days are cleared by GitHub.
https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy

Pip checks if a newer version of the dependency is available on PyPI and
use it instead of the cache when possible.
https://pip.pypa.io/en/stable/cli/pip_install/#satisfying-requirements

The linkcheck run was excluded as the cache would only grow. It’s a
scheduled build occurring once a week, caching wouldn’t be as
beneficial.