Allow testing against mongomock
ArchLinux and maybe other Linux distributions have interest in testing
their package of the project, in an environment where MongoDB isn’t
available.

Support for that environment is provided on a best-effort basis. If this
patch ends up being more trouble than it’s worth, please revert it.
Otherwise, let’s be helpful to others. :)