Export types per PEP 561
Basic typing support was added recently, exposing return types for
`create()` and `build()`:
https://github.com/FactoryBoy/factory_boy/pull/1059

However, because this package does not expose these types, consumers of
`factory-boy` cannot leverage the annotations!

Let's ensure that the package exports its types. It's okay for only a
subset of methods to be typed! (In other words, it's still compliant
with PEP 561 to have partial coverage).